+++ CHANGELOG.md
@@ -1,6 +1,23 @@
 # Changelog
 
 All notable changes to the Tarevity project will be documented in this file.
+## [1.0.2] - 2025-03-26
+
+### Added
+- Integrate todos into calendar and enhance todo list with date filtering 
+
+### Changed
+- useTodoFilters, calendar and TodoList for implementation
+
+### Fixed
+- Bugs with date filtering, clicking on day 31 in the calendar would redirect to day 30 due to timezone issues
+
+### Security
+- ...
+
+### Removed
+- ...
+
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).